requirements-api.txt
# Structure Analysis API (FastAPI + model)
# Install: pip install -r requirements-api.txt

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0

# Model (same as structural_damage_model)
torch>=2.0.0
torchvision
transformers>=4.40.0
accelerate
Pillow
